+++ IronSource/src/main/java/com/applovin/mediation/adapters/IronSourceMediationAdapter.java
@@ -30,7 +30,6 @@ import com.ironsource.mediationsdk.demandOnly.ISDemandOnlyInterstitialListener;
 import com.ironsource.mediationsdk.demandOnly.ISDemandOnlyRewardedVideoListener;
 import com.ironsource.mediationsdk.logger.IronSourceError;
 import com.ironsource.mediationsdk.utils.IronSourceConstants;
-import com.ironsource.mediationsdk.utils.IronSourceUtils;
 import com.unity3d.ironsourceads.AdSize;
 import com.unity3d.ironsourceads.InitListener;
 import com.unity3d.ironsourceads.InitRequest;
@@ -103,10 +102,10 @@ public class IronSourceMediationAdapter
             if ( isDoNotSell != null )
             {
                 // NOTE: `setMetaData` must be called _before_ initializing their SDK
-                IronSource.setMetaData( "do_not_sell", Boolean.toString( isDoNotSell ) );
+                IronSourceAds.setMetaData( "do_not_sell", Boolean.toString( isDoNotSell ) );
             }
 
-            IronSource.setAdaptersDebug( parameters.isTesting() );
+            IronSourceAds.enableDebugMode( parameters.isTesting() );
             IronSource.setISDemandOnlyInterstitialListener( ROUTER );
             IronSource.setISDemandOnlyRewardedVideoListener( ROUTER );
 
@@ -140,7 +139,7 @@ public class IronSourceMediationAdapter
     @Override
     public String getSdkVersion()
     {
-        return IronSourceUtils.getSDKVersion();
+        return IronSourceAds.getSdkVersion();
     }
 
     @Override
@@ -254,7 +253,9 @@ public class IronSourceMediationAdapter
             if ( biddingInterstitialAd == null || !biddingInterstitialAd.isReadyToShow() )
             {
                 log( "Unable to show ironSource interstitial - ad is not ready for instance ID: " + instanceId );
-                listener.onInterstitialAdDisplayFailed( new MaxAdapterError( MaxAdapterError.AD_DISPLAY_FAILED, 0, "Interstitial ad not ready" ) );
+                listener.onInterstitialAdDisplayFailed( new MaxAdapterError( MaxAdapterError.AD_DISPLAY_FAILED,
+                                                                             MaxAdapterError.AD_NOT_READY.getCode(),
+                                                                             MaxAdapterError.AD_NOT_READY.getMessage() ) );
                 return;
             }
 
@@ -275,7 +276,9 @@ public class IronSourceMediationAdapter
             if ( !IronSource.isISDemandOnlyInterstitialReady( instanceId ) )
             {
                 log( "Unable to show ironSource interstitial - no ad loaded for instance ID: " + instanceId );
-                ROUTER.onAdDisplayFailed( IronSourceRouter.getInterstitialRouterIdentifier( instanceId ), new MaxAdapterError( MaxAdapterError.AD_DISPLAY_FAILED, 0, "Interstitial ad not ready" ) );
+                ROUTER.onAdDisplayFailed( IronSourceRouter.getInterstitialRouterIdentifier( instanceId ), new MaxAdapterError( MaxAdapterError.AD_DISPLAY_FAILED,
+                                                                                                                               MaxAdapterError.AD_NOT_READY.getCode(),
+                                                                                                                               MaxAdapterError.AD_NOT_READY.getMessage() ) );
                 return;
             }
 
@@ -337,7 +340,9 @@ public class IronSourceMediationAdapter
             if ( biddingRewardedAd == null || !biddingRewardedAd.isReadyToShow() )
             {
                 log( "Unable to show ironSource rewarded - ad is not ready for instance ID: " + instanceId );
-                listener.onRewardedAdDisplayFailed( new MaxAdapterError( MaxAdapterError.AD_DISPLAY_FAILED, 0, "Rewarded ad not ready" ) );
+                listener.onRewardedAdDisplayFailed( new MaxAdapterError( MaxAdapterError.AD_DISPLAY_FAILED,
+                                                                         MaxAdapterError.AD_NOT_READY.getCode(),
+                                                                         MaxAdapterError.AD_NOT_READY.getMessage() ) );
                 return;
             }
 
@@ -361,7 +366,9 @@ public class IronSourceMediationAdapter
             if ( !IronSource.isISDemandOnlyRewardedVideoAvailable( instanceId ) )
             {
                 log( "Unable to show ironSource rewarded - no ad loaded..." );
-                ROUTER.onAdDisplayFailed( IronSourceRouter.getRewardedVideoRouterIdentifier( instanceId ), new MaxAdapterError( MaxAdapterError.AD_DISPLAY_FAILED, 0, "Rewarded ad not ready" ) );
+                ROUTER.onAdDisplayFailed( IronSourceRouter.getRewardedVideoRouterIdentifier( instanceId ), new MaxAdapterError( MaxAdapterError.AD_DISPLAY_FAILED,
+                                                                                                                                MaxAdapterError.AD_NOT_READY.getCode(),
+                                                                                                                                MaxAdapterError.AD_NOT_READY.getMessage() ) );
                 return;
             }
 
@@ -430,7 +437,7 @@ public class IronSourceMediationAdapter
         Boolean hasUserConsent = parameters.hasUserConsent();
         if ( hasUserConsent != null )
         {
-            IronSource.setConsent( hasUserConsent );
+            IronSourceAds.setConsent( hasUserConsent );
         }
     }
 
@@ -691,7 +698,9 @@ public class IronSourceMediationAdapter
         @Override
         public void onInterstitialAdShowFailed(final String instanceId, final IronSourceError ironSourceError)
         {
-            MaxAdapterError adapterError = new MaxAdapterError( MaxAdapterError.AD_DISPLAY_FAILED, ironSourceError.getErrorCode(), ironSourceError.getErrorMessage() );
+            MaxAdapterError adapterError = new MaxAdapterError( MaxAdapterError.AD_DISPLAY_FAILED,
+                                                                ironSourceError.getErrorCode(),
+                                                                ironSourceError.getErrorMessage() );
             log( "Interstitial ad failed to show for instance ID: " + instanceId + " with error: " + adapterError );
             onAdDisplayFailed( IronSourceRouter.getInterstitialRouterIdentifier( instanceId ), adapterError );
         }
@@ -736,7 +745,9 @@ public class IronSourceMediationAdapter
         @Override
         public void onRewardedVideoAdShowFailed(final String instanceId, final IronSourceError ironSourceError)
         {
-            MaxAdapterError adapterError = new MaxAdapterError( MaxAdapterError.AD_DISPLAY_FAILED, ironSourceError.getErrorCode(), ironSourceError.getErrorMessage() );
+            MaxAdapterError adapterError = new MaxAdapterError( MaxAdapterError.AD_DISPLAY_FAILED,
+                                                                ironSourceError.getErrorCode(),
+                                                                ironSourceError.getErrorMessage() );
             log( "Rewarded ad failed to show for instance ID: " + instanceId + " with error: " + adapterError );
             onAdDisplayFailed( IronSourceRouter.getRewardedVideoRouterIdentifier( instanceId ), adapterError );
         }
@@ -774,12 +785,12 @@ public class IronSourceMediationAdapter
 
         private static String getInterstitialRouterIdentifier(final String instanceId)
         {
-            return instanceId + "-" + IronSource.AD_UNIT.INTERSTITIAL;
+            return instanceId + '-' + IronSource.AD_UNIT.INTERSTITIAL;
         }
 
         private static String getRewardedVideoRouterIdentifier(final String instanceId)
         {
-            return instanceId + "-" + IronSource.AD_UNIT.REWARDED_VIDEO;
+            return instanceId + '-' + IronSource.AD_UNIT.REWARDED_VIDEO;
         }
     }
 
@@ -876,7 +887,9 @@ public class IronSourceMediationAdapter
         @Override
         public void onInterstitialAdFailedToShow(@NonNull final InterstitialAd ad, @NonNull final IronSourceError ironSourceError)
         {
-            MaxAdapterError adapterError = new MaxAdapterError( MaxAdapterError.AD_DISPLAY_FAILED, ironSourceError.getErrorCode(), ironSourceError.getErrorMessage() );
+            MaxAdapterError adapterError = new MaxAdapterError( MaxAdapterError.AD_DISPLAY_FAILED,
+                                                                ironSourceError.getErrorCode(),
+                                                                ironSourceError.getErrorMessage() );
             log( "Interstitial ad failed to show for bidding instance ID: " + ad.getAdInfo().getInstanceId() + " with error: " + adapterError );
             listener.onInterstitialAdDisplayFailed( adapterError, createExtraInfo( ad ) );
         }
@@ -953,7 +966,9 @@ public class IronSourceMediationAdapter
         @Override
         public void onRewardedAdFailedToShow(@NonNull final RewardedAd ad, @NonNull final IronSourceError ironSourceError)
         {
-            MaxAdapterError adapterError = new MaxAdapterError( MaxAdapterError.AD_DISPLAY_FAILED, ironSourceError.getErrorCode(), ironSourceError.getErrorMessage() );
+            MaxAdapterError adapterError = new MaxAdapterError( MaxAdapterError.AD_DISPLAY_FAILED,
+                                                                ironSourceError.getErrorCode(),
+                                                                ironSourceError.getErrorMessage() );
             log( "Rewarded ad failed to show for bidding instance ID: " + ad.getAdInfo().getInstanceId() + " with error: " + adapterError );
             listener.onRewardedAdDisplayFailed( adapterError, createExtraInfo( ad ) );
         }